添加断点下载功能

+++ wdownloadtool/src/main/java/com/app/wpf/wdownloadtool/DownloadImpl/DownComplexInfo.java
@@ -7,6 +7,6 @@ package com.app.wpf.wdownloadtool.DownloadImpl;
 
 public abstract class DownComplexInfo extends WDownloadToolResult {
     public abstract void downloadPercent(int percent);
-    public abstract void downloadDetailed(int downSize);
+    public abstract void downloadDetailed(long downSize);
     public abstract void downloadSpeed(double speed);
 }
